Create Emulation module; add userAgent arbitrary
The userAgent's nonConforming behaviour is added to the property,
allowing us to produce another error and validate it.
